+++ src/utils/externalAPIs.ts
@@ -15,7 +15,7 @@ import { ext } from '../extensionVariables';
     status: string,
     statusText: string,
     headers: object,
-    data: object,
+    data?: object,
     request: {
         url: string,
         method: string,
@@ -26,6 +26,14 @@ import { ext } from '../extensionVariables';
 
  }
 
+/**
+ * wraps/expands AxiosRequestConfig
+ *  - used to easily set the http agent setting for allowing self-signed certs
+ */
+export interface HttpRequest extends AxiosRequestConfig {
+    rejectUnauthorized?: boolean,
+}
+
 
 /**
  * calls external HTTP APIs based on axsios.request parameters
@@ -33,31 +41,34 @@ import { ext } from '../extensionVariables';
  * 
  * @param req AxiosRequestConfig options
  */
-export async function makeRequest(req: AxiosRequestConfig) {
+export async function makeRequest(req: HttpRequest) {
 
     logger.debug('external http pre-Opts', JSON.stringify(req));
 
-    // const httpsAgent = new https.Agent({ rejectUnauthorized: false });
+    // add option to allow self-signed cert
+    if (req.rejectUnauthorized) {
+        req.httpsAgent = new https.Agent({ rejectUnauthorized: false });
+    }
+
+    // get port from URL
     
     // rewrite req object with defaults
-    req = {
-        url: req.url,
-        method: req['method'] || 'GET',
-        data: req['data'] || null,
-        httpsAgent: new https.Agent({
-            rejectUnauthorized: false
-        }),
-        validateStatus: () => true
-    };
+    // req = {
+    //     url: req.url,
+    //     method: req['method'] || 'GET',
+    //     data: req['data'] || null,
+    //     httpsAgent: new https.Agent({
+    //         rejectUnauthorized: false
+    //     }),
+    //     validateStatus: () => true
+    // };
 
     logger.debug('external http defaults-Opts', JSON.stringify(req));
 
     const resp = await axios.request(req)
     .then( resp => {
-        // logger.debug('buuuug');
-        /**
-         * only return the things we want/need
-         */
+
+        // only return the stuff we want
         return {
             data: resp.data,
             headers: resp.headers,
@@ -68,41 +79,25 @@ export async function makeRequest(req: AxiosRequestConfig) {
                 method: resp.request.method,
                 headers: resp.request._headers,
                 protocol: resp.config.httpsAgent.protocol,
-                data: resp.data
+                // data: resp.data
             }
         };
-        // return resp;
     })
-    .catch(function (error) {
-        // debugger;
-        if (error.response) {
-            // The request was made and the server responded with a status code
-            // that falls out of the range of 2xx
-            // console.error(error.response.data);
-            // console.error(error.response.status);
-            // console.error(error.response.headers);
-
-            const status = error.response.status;
-            const message = error.response.data.message;
-
-            vscode.window.showErrorMessage(`HTTP_FAILURE: ${status} - ${message}`);
-            console.error(`HTTP_FAILURE: ${status} - ${message} - ${JSON.stringify(error.response.data)}`);
-            throw new Error(`HTTP_FAILURE: ${status} - ${message}`);
-        // } else if (error.request) {
-        //   // The request was made but no response was received
-        //   // `error.request` is an instance of XMLHttpRequest in the browser and an instance of
-        //   // http.ClientRequest in node.js
-        //   logger.debug('AuthHttpErrorRequest', error.request);
-        } else if (error.code && error.message){
-            // console.error('HTTP-response-error:', error);
-            console.error(`HTTP_response_error: ${error.code} - ${error.message}`);
-            vscode.window.showErrorMessage(`${error.code} - ${error.message}`);
-        } else {
-            // Something happened in setting up the request that triggered an Error
-            console.error('AuthHttpError', error.message);
-        }
-        // console.error('AuthHttpConfigError',error.config);
-        console.error('AuthHttpFULLError',error);
+    .catch( err => {
+        
+        // return all error information also
+        return {
+            data: err.response.data,
+            headers: err.response.headers,
+            status: err.response.status,
+            statusText: err.response.statusText,
+            request: {
+                url: err.config.url,
+                method: err.request.method,
+                headers: err.request._headers,
+                protocol: err.config.httpsAgent.protocol,
+            }
+        };
     });
 
     return resp;
@@ -176,130 +171,93 @@ export function callHTTPS(opts: object, payload: object = {}): Promise<any> {
     });
 };
 
-export function callHTTPSsync(opts: object, payload: object = {}) {
+// export function callHTTPSsync(opts: object, payload: object = {}) {
 
 
-    logger.debug('callHTTPS---OPTS: ' + JSON.stringify(opts));
-    // logger.debug('callHTTTS---payload: ' + JSON.stringify(payload));
+//     logger.debug('callHTTPS---OPTS: ' + JSON.stringify(opts));
+//     // logger.debug('callHTTTS---payload: ' + JSON.stringify(payload));
     
-    // logger.debug('Bout to call API token request')
-    // return new Promise((resolve, reject) => {
-        const req = request(opts, (res) => {
-            const buffer: any = [];
-            res.setEncoding('utf8');
-            // logger.debug('Sending::: ' )
-            res.on('data', (data) => {
-                buffer.push(data);
-            });
-            res.on('end', () => {
-                let body = buffer.join('');
-                body = body || '{}';
+//     // logger.debug('Bout to call API token request')
+//     // return new Promise((resolve, reject) => {
+//         const req = request(opts, (res) => {
+//             const buffer: any = [];
+//             res.setEncoding('utf8');
+//             // logger.debug('Sending::: ' )
+//             res.on('data', (data) => {
+//                 buffer.push(data);
+//             });
+//             res.on('end', () => {
+//                 let body = buffer.join('');
+//                 body = body || '{}';
 
-                // try {
-                //     body = JSON.parse(body);
-                // } catch (e) {
-                //     return reject(new Error(`Invalid response object ${opts}`));
-                // };
+//                 // try {
+//                 //     body = JSON.parse(body);
+//                 // } catch (e) {
+//                 //     return reject(new Error(`Invalid response object ${opts}`));
+//                 // };
                 
-                 // configure global logging parameters
-                logger.debug('callHTTPS***STATUS: ' + res.statusCode);
-                logger.debug('callHTTPS***HEADERS: ' + JSON.stringify(res.headers));
-                // logger.debug('callHTTPS***BODY: ' + JSON.stringify(body));
-                // logger.debug('callHTTPS***BODY: ' + body);
+//                  // configure global logging parameters
+//                 logger.debug('callHTTPS***STATUS: ' + res.statusCode);
+//                 logger.debug('callHTTPS***HEADERS: ' + JSON.stringify(res.headers));
+//                 // logger.debug('callHTTPS***BODY: ' + JSON.stringify(body));
+//                 // logger.debug('callHTTPS***BODY: ' + body);
 
-                // if (res.statusCode == 401) {
-                //     logger.debug(`GOT 401!!!!!`)
-                // }
+//                 // if (res.statusCode == 401) {
+//                 //     logger.debug(`GOT 401!!!!!`)
+//                 // }
                 
-                const goodResp: Array<number> = [200, 201, 202];
-                // was trying to check against array above with arr.includes or arr.indexOf
-                if (res.statusCode === 200 ) {
-                    logger.debug(`CAUGHT 200: `);
-                    return {
-                        status: res.statusCode,
-                        headers: res.headers,
-                        body
-                    };
-                } else {
+//                 const goodResp: Array<number> = [200, 201, 202];
+//                 // was trying to check against array above with arr.includes or arr.indexOf
+//                 if (res.statusCode === 200 ) {
+//                     logger.debug(`CAUGHT 200: `);
+//                     return {
+//                         status: res.statusCode,
+//                         headers: res.headers,
+//                         body
+//                     };
+//                 } else {
 
-                    console.error(`HTTP FAILURE: ${res.statusCode} - ${res.statusMessage}`);
-                    new Error(`HTTP - ${res.statusCode} - ${res.statusMessage}`);
-                }
-            });
-        });
-        // logger.debug(`req in callHTTPS: ${req}`);
+//                     console.error(`HTTP FAILURE: ${res.statusCode} - ${res.statusMessage}`);
+//                     new Error(`HTTP - ${res.statusCode} - ${res.statusMessage}`);
+//                 }
+//             });
+//         });
+//         // logger.debug(`req in callHTTPS: ${req}`);
 
-        req.on('error', (e) => {
-            // might need to stringify combOpts for proper log output
-            // reject(new Error(`${opts}:${e.message}`));
-            new Error(`${opts}:${e.message}`);
-        });
+//         req.on('error', (e) => {
+//             // might need to stringify combOpts for proper log output
+//             // reject(new Error(`${opts}:${e.message}`));
+//             new Error(`${opts}:${e.message}`);
+//         });
 
-        // if a payload was passed in, post it!
-        if (payload) {
-            req.write(JSON.stringify(payload));
-        }
-        req.end();
-    // });
-};
-
-
-
-// const getAuthToken = async (host: string, username: string, password: string) => callHTTPS({
-//     host,
-//     path: '/mgmt/shared/authn/login',
-//     method: 'POST',
-// }, 
-// { 
-//     username,
-//     password
-// })
-// .then( response => {
-//     if (response.status === 200) {
-//         return { 
-//             host: host, 
-//             token: response.body.token.token 
+//         // if a payload was passed in, post it!
+//         if (payload) {
+//             req.write(JSON.stringify(payload));
 //         }
-//     } else {
-//         // clear cached password for this device
-//         // ext.keyTar.deletePassword(
-//         //     'f5Hosts',
-//         //     `${username}@${host}`
-//         //     )
-//         //     throw new Error(`error from getAuthTokenNOT200: ${response}`);
-//     }
-    
-//     // if (response.status != 200) {
-//     //     // clear cached password for this device
-//     //     ext.keyTar.deletePassword(
-//     //         'f5Hosts',
-//     //         `${username}@${host}`
-//     //         )
-//     //     throw new Error(`error from getAuthTokenNOT200: ${response}`);
-//     // }
-//     // return { 
-//     //     host: host, 
-//     //     token: response.body.token.token 
-//     // };
+//         req.end();
+//     // });
+// };
+
+
+
+
+
+// const callHTTP = (method: string, host: string, path: string, token: string, payload: object = {}) => callHTTPS(
+//     {
+//         method,
+//         host,
+//         path,
+//         headers: {
+//             'Content-Type': 'application/json',
+//             'X-F5-Auth-Token': token
+//         }
+//     },
+//     payload
+// )
+// .then( response => {
+//     // logger.debug('response from callHTTP: ' + JSON.stringify(response));
+//     // Promise.resolve(value.body.token);
+//     return response;
 // });
-
-
-const callHTTP = (method: string, host: string, path: string, token: string, payload: object = {}) => callHTTPS(
-    {
-        method,
-        host,
-        path,
-        headers: {
-            'Content-Type': 'application/json',
-            'X-F5-Auth-Token': token
-        }
-    },
-    payload
-)
-.then( response => {
-    // logger.debug('response from callHTTP: ' + JSON.stringify(response));
-    // Promise.resolve(value.body.token);
-    return response;
-});